_legacy_mvp/server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
python-multipart>=0.0.9
google-genai>=1.53.0
python-dotenv>=1.0.0
pydantic>=2.12.5
pydub>=0.25.1
psycopg2-binary>=2.9.9
websocket-client>=1.6.0
requests>=2.31.0

# 测试依赖
pytest>=8.0.0
pytest-json-report>=1.5.0
